jpa integration with MySQL + impl ProductRepository + Add jpa entity

+++ src/main/java/dev/jclp/demo/product_command/entities/Product.java
@@ -2,6 +2,7 @@ package dev.jclp.demo.product_command.entities;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -15,7 +16,7 @@ public class Product {
     private String name;
 
     @Column(precision = 10, scale = 2)
-    private Double price;
+    private BigDecimal price;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
@@ -26,7 +27,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(String name, Double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -58,11 +59,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
